+++ src/main/java/me/pajic/rearm/CompatFlags.java
@@ -2,14 +2,15 @@ package me.pajic.rearm;
 
 import net.fabricmc.loader.api.FabricLoader;
 
-import java.util.List;
+import java.util.Map;
 
 public class CompatFlags {
-    public static final List<String> ENCHANTMENT_DESCRIPTION_MODS = List.of(
-            "enchdesc",
-            "idwtialsimmoedm"
+    public static final Map<String, String> ENCHANTMENT_DESCRIPTION_MODS = Map.of(
+            "enchdesc", "desc",
+            "idwtialsimmoedm", "desc",
+            "item-descriptions", "description"
     );
     public static boolean HMI_LOADED = FabricLoader.getInstance().isModLoaded("hold-my-items");
     public static boolean SHIELD_LIB_LOADED = FabricLoader.getInstance().isModLoaded("fabricshieldlib") || FabricLoader.getInstance().isModLoaded("shieldlib");
-    public static boolean ENCHDESC_MOD_LOADED = ENCHANTMENT_DESCRIPTION_MODS.stream().anyMatch(mod -> FabricLoader.getInstance().isModLoaded(mod));
+    public static boolean ENCHDESC_MOD_LOADED = ENCHANTMENT_DESCRIPTION_MODS.keySet().stream().anyMatch(mod -> FabricLoader.getInstance().isModLoaded(mod));
 }