+++ src/main/java/me/pajic/rearm/mixson/ResourceModifications.java
@@ -11,7 +11,6 @@ import net.minecraft.client.Minecraft;
 import net.ramixin.mixson.debug.DebugMode;
 import net.ramixin.mixson.inline.EventContext;
 import net.ramixin.mixson.inline.Mixson;
-import net.ramixin.mixson.util.MixsonUtil;
 
 import java.util.List;
 
@@ -293,83 +292,37 @@ public class ResourceModifications {
                 },
                 true
         );
-        Mixson.registerEvent(
-                Mixson.DEFAULT_PRIORITY,
-                rl -> rl.toString().startsWith("rearm:lang/"),
-                "Apply enchantment name overrides",
-                context -> {
-                    if (context.getResourceId().getPath().contains(Minecraft.getInstance().getLanguageManager().getSelected())) {
-                        if (Main.CONFIG.protection.elementalProtection.get()) {
-                            context.getFile().getAsJsonObject().addProperty(
-                                    "enchantment.minecraft.fire_protection",
-                                    context.getFile().getAsJsonObject().get("enchantment.minecraft.fire_protection.override").getAsString()
-                            );
-                        }
-                        if (Main.CONFIG.protection.meleeProtection.get()) {
-                            context.getFile().getAsJsonObject().addProperty(
-                                    "enchantment.minecraft.protection",
-                                    context.getFile().getAsJsonObject().get("enchantment.minecraft.protection.override").getAsString()
-                            );
-                        }
-                    }
-                },
-                true
-        );
         if (CompatFlags.ENCHDESC_MOD_LOADED) {
-            CompatFlags.ENCHANTMENT_DESCRIPTION_MODS.forEach(mod -> Mixson.registerEvent(
+            CompatFlags.ENCHANTMENT_DESCRIPTION_MODS.forEach((mod, suffix) -> Mixson.registerEvent(
                     Mixson.DEFAULT_PRIORITY,
-                    MixsonUtil.getLocatorFromString(mod + ":lang/*"),
+                    rl -> rl.toString().startsWith(mod.replace('-', '_') + ":lang/"),
                     "Apply enchantment description overrides for " + mod,
                     context -> {
                         if (context.getResourceId().getPath().contains(Minecraft.getInstance().getLanguageManager().getSelected())) {
                             try {
                                 if (Main.CONFIG.bow.improvedMultishot.get()) {
-                                    context.getFile().getAsJsonObject().addProperty(
-                                            "enchantment.minecraft.multishot.desc",
-                                            context.getFile().getAsJsonObject().get("enchantment.minecraft.multishot.desc.override").getAsString()
-                                    );
+                                    context.getFile().getAsJsonObject().remove("enchantment.minecraft.multishot." + suffix);
                                 }
                                 if (Main.CONFIG.crossbow.improvedPiercing.get()) {
-                                    context.getFile().getAsJsonObject().addProperty(
-                                            "enchantment.minecraft.piercing.desc",
-                                            context.getFile().getAsJsonObject().get("enchantment.minecraft.piercing.desc.override").getAsString()
-                                    );
+                                    context.getFile().getAsJsonObject().remove("enchantment.minecraft.piercing." + suffix);
                                 }
                                 if (Main.CONFIG.sword.improvedSweepingEdge.get()) {
-                                    context.getFile().getAsJsonObject().addProperty(
-                                            "enchantment.minecraft.sweeping_edge.desc",
-                                            context.getFile().getAsJsonObject().get("enchantment.minecraft.sweeping_edge.desc.override").getAsString()
-                                    );
+                                    context.getFile().getAsJsonObject().remove("enchantment.minecraft.sweeping_edge." + suffix);
                                 }
                                 if (Main.CONFIG.protection.elementalProtection.get()) {
-                                    context.getFile().getAsJsonObject().addProperty(
-                                            "enchantment.minecraft.fire_protection.desc",
-                                            context.getFile().getAsJsonObject().get("enchantment.minecraft.fire_protection.desc.override").getAsString()
-                                    );
+                                    context.getFile().getAsJsonObject().remove("enchantment.minecraft.fire_protection." + suffix);
                                 }
                                 if (Main.CONFIG.protection.meleeProtection.get()) {
-                                    context.getFile().getAsJsonObject().addProperty(
-                                            "enchantment.minecraft.protection.desc",
-                                            context.getFile().getAsJsonObject().get("enchantment.minecraft.protection.desc.override").getAsString()
-                                    );
+                                    context.getFile().getAsJsonObject().remove("enchantment.minecraft.protection." + suffix);
                                 }
                                 if (Main.CONFIG.tweaks.infinityFix.get()) {
-                                    context.getFile().getAsJsonObject().addProperty(
-                                            "enchantment.minecraft.infinity.desc",
-                                            context.getFile().getAsJsonObject().get("enchantment.minecraft.infinity.desc.override").getAsString()
-                                    );
+                                    context.getFile().getAsJsonObject().remove("enchantment.minecraft.infinity." + suffix);
                                 }
                                 if (Main.CONFIG.crossbow.acceptPower.get()) {
-                                    context.getFile().getAsJsonObject().addProperty(
-                                            "enchantment.minecraft.power.desc",
-                                            context.getFile().getAsJsonObject().get("enchantment.minecraft.power.desc.override").getAsString()
-                                    );
+                                    context.getFile().getAsJsonObject().remove("enchantment.minecraft.power." + suffix);
                                 }
                                 if (context.getFile().getAsJsonObject().has("enchdesc.activate.message")) {
-                                    context.getFile().getAsJsonObject().addProperty(
-                                            "enchdesc.activate.message",
-                                            context.getFile().getAsJsonObject().get("enchdesc.activate.message.override").getAsString()
-                                    );
+                                    context.getFile().getAsJsonObject().remove("enchdesc.activate.message");
                                 }
                             } catch (NullPointerException ignored) {}
                         }